LayoutTests/imported/w3c/web-platform-tests/html/syntax/charset/xhr-expected.txt


FAIL Check after-1kb.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
PASS Check after-bogus-after-1kb.html
FAIL Check after-bogus.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
PASS Check after-head-after-1kb-crlf.html
PASS Check after-head-after-1kb.html
FAIL Check after-head-in-1kb-crlf.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
FAIL Check after-head-in-1kb.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
FAIL Check baseline.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
PASS Check document-write.html
PASS Check in-comment.html
PASS Check in-noscript-after-template-after-1kb.html
FAIL Check in-object.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
PASS Check in-script.html
PASS Check in-style.html
FAIL Check in-svg.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
PASS Check in-svg-in-cdata.html
PASS Check in-template-after-1kb.html
FAIL Check in-template.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
PASS Check in-title.html
FAIL Check ncr.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
FAIL Check non-ascii-in-comment-before.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
FAIL Check non-ascii-in-title-before.html assert_equals: Check characterSet expected "windows-1251" but got "UTF-8"
